+++ routes/index.js
@@ -1,4 +1,13 @@
-const publicRoute = require("./privateRoute");
-const privateRoute = require("./privateRoute");
+const publicRoute = require("./usersRoute");
+const usersRoute = require("./usersRoute");
+const postsRoute = require("./postsRoute");
+const birthChartRoute = require("./birthchartRoute");
+const friendsRoute = require("./friendsRoute");
 
-module.exports = { publicRoute, privateRoute };
+module.exports = {
+  publicRoute,
+  usersRoute,
+  postsRoute,
+  birthChartRoute,
+  friendsRoute,
+};
